fix delete eip bug

+++ src/main/java/com/inspur/eip/service/EipDaoService.java
@@ -153,12 +153,13 @@ public class EipDaoService {
                 msg = "Failed to delete floating ip, floatingIpId:" + eipEntity.getFloatingIpId();
                 log.error(msg);
             }
-
-            ActionResponse delV6Ret = eipV6DaoService.deleteEipV6(eipEntity.getEipV6Id(), token);
-            if (!delV6Ret.isSuccess()) {
-                msg = "Faild to delete ipv6 address.";
-                log.error(msg);
-                return ActionResponse.actionFailed(msg, HttpStatus.SC_INTERNAL_SERVER_ERROR);
+            if(eipEntity.getEipV6Id() != null){
+                ActionResponse delV6Ret = eipV6DaoService.deleteEipV6(eipEntity.getEipV6Id(), token);
+                if (!delV6Ret.isSuccess()) {
+                    msg = "Faild to delete ipv6 address.";
+                    log.error(msg);
+                    return ActionResponse.actionFailed(msg, HttpStatus.SC_INTERNAL_SERVER_ERROR);
+                }
             }
 
             eipEntity.setIsDelete(1);
